requirements.txt
typer
pyyaml
tqdm
duckdb
pandas
pyarrow
zstandard
requests
